requirements.txt
asgiref==3.8.1
Django==5.0.3
django-jazzmin==2.6.1
djangorestframework==3.15.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
